+++ inst/scrape.R
@@ -1,0 +1,3 @@
+library(coronavirus)
+googlesheets4::sheets_deauth()
+crawl_coronavirus()